+++ server/constants.js
@@ -1,0 +1,4 @@
+module.exports = {
+    SPEED_UNKNOWN: -1,
+    ALTITUDE_UNKNOWN: -1,
+}